Parte Inicial de salvar movimentações e algumas correções na conta Busca dos feriados Listagem padrão de usuários para auxiliar nos testes

+++ BancoChu/Entidades/Contas.cs
@@ -6,7 +6,7 @@
         public int Agencia { get; set; }
         public string CpfCnpj { get; set; }
         public string Nome { get; set; }
-        public double Saldo { get; set; }
+        public decimal Saldo { get; set; }
         public string Email { get; set; }
         public DateTime DataCriacao { get; set; }
         public DateTime DataAlterado { get; set; }
